Create abstracted DataResource.
